changed tag type to NSInteger
git-svn-id: svn://orca.physics.unc.edu/Orca@11747 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/OrcaObject/OrcaObject.h
@@ -54,7 +54,7 @@ extern NSString* ORObjArrayPtrPBType;
     @private
         //internal flags
         BOOL		aWake;
-        int			tag; //used by subclasses for identification
+        NSInteger   tag; //used by subclasses for identification
         unsigned long uniqueIdNumber;
 }
 
@@ -100,10 +100,10 @@ extern NSString* ORObjArrayPtrPBType;
 - (void)		setSkipConnectionDraw:(BOOL)state;
 - (NSMutableArray*) 	children;
 - (NSMutableArray*) 	familyList;
-- (int)         stationNumber;
-- (int)			tag;
-- (void)		setTag:(int)aTag;
-- (int)         tagBase;
+- (NSInteger)         stationNumber;
+- (NSInteger)	tag;
+- (void)		setTag:(NSInteger)aTag;
+- (NSInteger)         tagBase;
 - (BOOL)        solitaryObject;
 - (BOOL)		solitaryInViewObject;
 - (void)        askForUniqueIDNumber;
